Chapter 7 - Develop reactive microservices

+++ recommendation-service/src/main/java/study/microservices/core/recommendation/services/RecommendationServiceImpl.java
@@ -1,19 +1,19 @@
 package study.microservices.core.recommendation.services;
 
+import static java.util.logging.Level.FINE;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 import study.api.core.recommendation.Recommendation;
 import study.api.core.recommendation.RecommendationService;
 import study.api.exceptions.InvalidInputException;
-import study.microservices.core.recommendation.persistence.RecommendationEntity;
 import study.microservices.core.recommendation.persistence.RecommendationRepository;
 import study.util.http.ServiceUtil;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @RestController
 public class RecommendationServiceImpl implements RecommendationService {
@@ -36,40 +36,48 @@ public class RecommendationServiceImpl implements RecommendationService {
   }
 
   @Override
-  public List<Recommendation> getRecommendations(int productId) {
+  public Flux<Recommendation> getRecommendations(int productId) {
     if (productId < 1) {
       throw new InvalidInputException("Invalid productId: " + productId);
     }
 
-    List<RecommendationEntity> entityList = repository.findByProductId(productId);
-    List<Recommendation> list = mapper.entityListToApiList(entityList);
-    list.forEach(e -> e.setServiceAddress(serviceUtil.getServiceAddress()));
+    LOG.info("Will get recommendations for product with id={}", productId);
 
-    LOG.debug("getRecommendations: response size: {}", list.size());
-
-    return list;
+    return repository.findByProductId(productId)
+        .log(LOG.getName(), FINE)
+        .map(mapper::entityToApi)
+        .map(this::setServiceAddress);
   }
 
   @Override
-  public Recommendation createRecommendation(Recommendation body) {
-    try {
-      RecommendationEntity savedEntity = repository.save(mapper.apiToEntity(body));
-
-      LOG.debug("createRecommendation: created a recommendation entity: {}/{}",
-          savedEntity.getProductId(), savedEntity.getRecommendationId());
-      return mapper.entityToApi(savedEntity);
-    } catch (DuplicateKeyException e) {
-      throw new InvalidInputException(
-          "Duplicate key, Product Id: " + body.getProductId() + ", Recommendation Id:"
-              + body.getRecommendationId());
+  public Mono<Recommendation> createRecommendation(Recommendation body) {
+    if (body.getProductId() < 1) {
+      throw new InvalidInputException("Invalid productId: " + body.getProductId());
     }
+
+    return repository.save(mapper.apiToEntity(body))
+        .log(LOG.getName(), FINE)
+        .onErrorMap(
+            DuplicateKeyException.class,
+            ex -> new InvalidInputException("Duplicate key, Product Id: " + body.getProductId()
+                + ", Recommendation Id:" + body.getRecommendationId()))
+        .map(mapper::entityToApi);
   }
 
   @Override
-  public void deleteRecommendations(int productId) {
+  public Mono<Void> deleteRecommendations(int productId) {
+    if (productId < 1) {
+      throw new InvalidInputException("Invalid productId: " + productId);
+    }
+
     LOG.debug(
         "deleteRecommendations: tries to delete recommendations for the product with productId: {}",
         productId);
-    repository.deleteAll(repository.findByProductId(productId));
+    return repository.deleteAll(repository.findByProductId(productId));
+  }
+
+  private Recommendation setServiceAddress(Recommendation r) {
+    r.setServiceAddress(serviceUtil.getServiceAddress());
+    return r;
   }
 }
